Change video processing function and imported data library in user interface
